+++ api/contest/upload.php
@@ -1,6 +1,6 @@
 <?php
     //? |-----------------------------------------------------------------------------------------------|
-    //? |  /api/contest/upload.php                                                                         |
+    //? |  /api/contest/upload.php                                                                      |
     //? |                                                                                               |
     //? |  Copyright (c) 2018-2019 Belikhun. All right reserved                                         |
     //? |  Licensed under the MIT License. See LICENSE in the project root for license information.     |