+++ index.php
@@ -75,19 +75,19 @@
         <!-- Load Library First -->
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/default.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/splash.css?v=<?php print VERSION; ?>" />
-        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/progressbar.css?v=<?php print VERSION; ?>" />
+        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/progressBar.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/button.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/input.css?v=<?php print VERSION; ?>" />
-        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/textview.css?v=<?php print VERSION; ?>" />
+        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/textView.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/table.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/switch.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/slider.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/navbar.css?v=<?php print VERSION; ?>" />
-        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/usersetting.css?v=<?php print VERSION; ?>" />
+        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/userSetting.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/menu.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/spinner.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/statusBar.css?v=<?php print VERSION; ?>" />
-        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/scrollbar.css?v=<?php print VERSION; ?>" />
+        <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/scrollBar.css?v=<?php print VERSION; ?>" />
         <!-- Page Style -->
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/core.css?v=<?php print VERSION; ?>" />
         <link rel="stylesheet" type="text/css" media="screen" href="/assets/css/dark.css?v=<?php print VERSION; ?>" />
@@ -147,7 +147,7 @@
                                 .join("")
                             : `<li>${e} >>> ${d}</li>`;
                         
-                        errorDetail.classList.add("textview", "small", "noBreakLine");
+                        errorDetail.classList.add("textView", "small", "noBreakLine");
                         errorDetail.style.flexDirection = "column";
                         errorDetail.innerHTML = errorDetailHtml;
 
@@ -215,7 +215,7 @@
             </span>
         </div>
 
-        <span id="user_settings" class="sound" data-soundtoggle="show">
+        <span id="userSettings" class="sound" data-soundtoggle="show">
 
             <div class="main">
                 <div class="container menu">
@@ -548,6 +548,9 @@
                                         <textarea id="problemEdit_desc" class="formField sound" placeholder="Nội dung" data-soundselectsoft required></textarea>
                                         <label for="problemEdit_desc">Nội dung</label>
                                     </div>
+                                    <div class="note info">
+                                        <span class="inner">Nếu tệp đính kèm là <b>pdf</b>, hệ thống sẽ hiển thị tệp đó trong cửa sổ xem đề bài</span>
+                                    </div>
                                     <div class="formGroup blue">
                                         <input id="problemEdit_attachment" type="file" class="formField sound" placeholder="Tệp đính kèm" data-soundselectsoft>
                                         <label for="problemEdit_attachment">Tệp đính kèm (tùy chọn)</label>
@@ -736,7 +739,7 @@
                         <t id="submitStatus">null</t>
                         <t id="submitFileName">null</t>
                         <div class="progressBar">
-                            <div class="bar" id="submitProgressBar"></div>
+                            <div class="bar" id="submitprogressBar"></div>
                             <t class="left" id="submitInfoProgress">0%</t>
                             <t class="right" id="submitInfoSize">00/00</t>
                         </div>